# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ydisks 闲鱼助手 API
  version: 1.0.0
  description: 版本化 HTTP API 的唯一契约源。
servers:
  - url: /
paths:
...
  "/api/v1/orders/{order_id}":
    delete:
      operationId: deleteApiV1OrdersByorder_id
      summary: deleteApiV1OrdersByorder_id
      parameters:
          - name: order_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: getApiV1OrdersByorder_id
      summary: getApiV1OrdersByorder_id
      parameters:
          - name: order_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: putApiV1OrdersByorder_id
      summary: putApiV1OrdersByorder_id
      parameters:
          - name: order_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/v1/orders/refresh":
    post:
      operationId: postApiV1OrdersRefresh
      summary: postApiV1OrdersRefresh
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRefreshRequest'
          multipart/form-data:
            schema: {$ref: '#/components/schemas/OrderRefreshRequest'}
      responses:
        '202':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderRefreshJobStartResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/orders/refresh/{job_id}":
    delete:
      operationId: deleteApiV1OrdersRefreshByjob_id
      summary: deleteApiV1OrdersRefreshByjob_id
      parameters:
          - name: job_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderRefreshJobCancelResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: getApiV1OrdersRefreshByjob_id
      summary: getApiV1OrdersRefreshByjob_id
      parameters:
          - name: job_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderRefreshJobStatusResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
      description: 统一响应的稳定字段；业务成功字段由各阶段 operation schema 收紧。
...
    OrderDTO:
      type: object
      required: [order_id, item_id, item_title, item_image, buyer_id, quantity, amount, order_status, status, cookie_id, created_at]
      properties:
        order_id: {type: string}
        item_id: {type: string}
        item_title: {type: string}
        item_image: {type: string}
        buyer_id: {type: string}
        spec_name: {type: string}
        spec_value: {type: string}
        quantity: {type: string}
        amount: {type: string}
        order_status: {type: string}
        status: {type: string}
        cookie_id: {type: string}
        is_bargain: {type: integer}
        system_shipped: {type: boolean}
        receiver_name: {type: string}
        receiver_phone: {type: string}
        receiver_address: {type: string}
        receiver_city: {type: string}
        created_at: {type: string}
        updated_at: {type: string}
...
    OrderDetailResponse:
      allOf:
        - {$ref: '#/components/schemas/OrderDTO'}
        - type: object
          required: [success, data]
          properties:
            success: {type: boolean}
            data: {$ref: '#/components/schemas/OrderDTO'}
    OrderUpdateRequest:
      type: object
      properties:
        order_status: {type: string}
        status: {type: string}
        item_id: {type: string}
        buyer_id: {type: string}
        spec_name: {type: string}
        spec_value: {type: string}
        quantity: {oneOf: [{type: string}, {type: number}]}
        amount: {oneOf: [{type: string}, {type: number}]}
        receiver_name: {type: string}
        receiver_phone: {type: string}
        receiver_address: {type: string}
        receiver_city: {type: string}
        chat_id: {type: string}
        system_shipped: {type: boolean}
        item_title: {type: string}
    OrderRefreshRequest:
      type: object
      properties:
        cookie_id: {type: string}
        status: {type: string}
...
    OrderRefreshResult:
      type: object
      required: [success]
      properties:
        success: {type: boolean}
        cookie_id: {type: string}
        stage: {type: string}
        message: {type: string}
        error: {type: string}
        discovered: {type: integer}
        updated: {type: integer}
        soft_deleted: {type: boolean}
        order_id: {type: string}
        old_status: {type: string}
        new_status: {type: string}
    OrderRefreshSummary:
      type: object
      required: [discovered, list_updated, soft_deleted, detail_total, total, updated, no_change, failed]
      properties:
        discovered: {type: integer}
        list_updated: {type: integer}
        soft_deleted: {type: integer}
        detail_total: {type: integer}
        total: {type: integer}
        updated: {type: integer}
        no_change: {type: integer}
        failed: {type: integer}
    OrderRefreshResponse:
      type: object
      required: [partial_failure, message, summary, results]
      properties:
        partial_failure: {type: boolean}
        message: {type: string}
        summary: {$ref: '#/components/schemas/OrderRefreshSummary'}
        results: {type: array, items: {$ref: '#/components/schemas/OrderRefreshResult'}}
    OrderRefreshJobStartResponse:
      type: object
      required: [success, job_id, status]
      properties:
        success: {type: boolean}
        job_id: {type: string}
        status: {type: string, enum: [queued, running]}
    OrderRefreshJobStatusResponse:
      type: object
      required: [success, job_id, status]
      properties:
        success: {type: boolean}
        job_id: {type: string}
        status: {type: string, enum: [queued, running, succeeded, failed, cancelled]}
        error_message: {type: string}
        result: {$ref: '#/components/schemas/OrderRefreshResponse'}
    OrderRefreshJobCancelResponse:
      type: object
      required: [success, job_id, status]
      properties:
        success: {type: boolean}
        job_id: {type: string}
        status: {type: string, enum: [cancelled]}
...
    OperationResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
        requires_relogin:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
      additionalProperties: true